services/jataware_georef/pyproject.toml
[tool.poetry]
name = "jataware-georef"
version = "0.1.0"
description = ""
authors = ["marshHawk4 <kyle@jataware.com>"]
readme = "README.md"
packages = [
         {include = "jataware_georef"},
         {include = "jataware_georef/**/*.py"}
 ]


[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
numpy = "^1.26.4"
rasterio = "^1.3.9"
uvicorn = "^0.23.2"
fastapi = "^0.110.0"
pydantic-settings = "^2.2.1"
pyproj = "^3.6.1"
httpx = "^0.26.0"
pillow = "^10.1.0"
shapely="2.0.1"
PyYAML="6.0.1"
boto3 = "^1.28.52"
albumentations="1.2.1"
segmentation-models-pytorch="0.3.0"
cdr_schemas = { git = "https://github.com/DARPA-CRITICALMAAS/cdr_schemas.git", tag = "v0.1.0+draft" }
torch = [
    # {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_aarch64.whl", platform = "linux", python = ">=3.10 <3.11", markers="platform_machine == \"arm64\""},
    {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl", platform = "linux", python = ">=3.10 <3.11", markers="platform_machine == \"amd64\""},
    {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1-cp310-none-macosx_11_0_arm64.whl", platform = "darwin", python = ">=3.10 <3.11", markers="platform_machine == \"arm64\""},
    {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1-cp310-none-macosx_10_9_x86_64.whl", platform = "darwin", python = ">=3.10 <3.11", markers="platform_machine == \"x86_64\""}
]


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
flake8-print = "^5.0.0"
autoflake = "^2.2.0"
uvicorn = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
